Rename item-list-entry-dropdown to dropdown-menu and use slot for actions
