--- client/src/elements/molecules/item-list-entry-dropdown.html
@@ -1,10 +1,0 @@
-<template>
-  <require from="./item-list-entry-dropdown.css"></require>
-  <button-tertiary class="item-list-entry-dropdown__button" icon="more" size="big" click.delegate="toggleOptionDropdown()"></button-tertiary>
-  <div ref="optionsDropdown" class="item-list-entry-dropdown__content q-text-small">
-    <a if.bind="!item.conf.active" click.delegate="activateItem()">${'item.activate' & t }</a>
-    <a if.bind="item.conf.active" click.delegate="deactivateItem()">${'item.deactivate' & t }</a>
-    <a click.delegate="blueprint()">${'item.blueprint' & t}</a>
-    <a if.bind="user.roles.includes('poweruser') || user.username === item.conf.createdBy" size="small" click.delegate="deleteItem()">${'item.delete' & t}</a>
-  </div>
-</template>
