+++ client/src/elements/molecules/item-list-entry.js
@@ -1,12 +1,32 @@
 import { bindable, inject } from "aurelia-framework";
 import ToolsInfo from "resources/ToolsInfo";
+import { Router } from "aurelia-router";
+import { Notification } from "aurelia-notification";
+import { I18N } from "aurelia-i18n";
+import { DialogService } from "aurelia-dialog";
+import { ConfirmDialog } from "dialogs/confirm-dialog.js";
+import User from "resources/User";
 
-@inject(ToolsInfo)
+@inject(ToolsInfo, Element, User, Router, Notification, I18N, DialogService)
 export class ItemListEntry {
   @bindable item;
 
-  constructor(toolsInfo) {
+  constructor(
+    toolsInfo,
+    element,
+    user,
+    router,
+    notification,
+    i18n,
+    dialogService
+  ) {
     this.toolsInfo = toolsInfo;
+    this.element = element;
+    this.user = user;
+    this.router = router;
+    this.notification = notification;
+    this.i18n = i18n;
+    this.dialogService = dialogService;
   }
 
   itemChanged() {
@@ -19,5 +39,45 @@ export class ItemListEntry {
         this.iconSvg = itemTool.icon;
       }
     });
+  }
+
+  activateItem() {
+    this.item.activate();
+  }
+
+  deactivateItem() {
+    this.item.deactivate();
+  }
+
+  async deleteItem() {
+    const openDialogResult = await this.dialogService.open({
+      viewModel: ConfirmDialog,
+      model: {
+        confirmQuestion: this.i18n.tr("item.questionDeleteItem"),
+        proceedText: this.i18n.tr("item.confirmDeleteItem"),
+        cancelText: this.i18n.tr("item.cancelDeleteItem")
+      }
+    });
+    const closeResult = await openDialogResult.closeResult;
+
+    if (!closeResult.wasCancelled) {
+      this.item.delete().then(() => {
+        this.element.addEventListener("transitionend", () => {
+          this.element.parentNode.removeChild(this.element);
+        });
+        this.element.style.transform = "scale(0)";
+      });
+    }
+  }
+
+  async blueprint() {
+    try {
+      await this.item.blueprint();
+      this.item.conf.title = this.i18n.tr("item.blueprintTitlePrefix");
+      let tool = this.item.conf.tool.replace(new RegExp("-", "g"), "_");
+      this.router.navigate(`/editor/${tool}/${this.item.conf._id}`);
+    } catch (e) {
+      this.notification.error("notification.failedToLoadItem");
+    }
   }
 }
